backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2

# Database
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.3

# Redis
redis[asyncio]==5.1.1

# HTTP client (for Ollama + RAGFlow)
httpx==0.27.2

# Clerk JWT verification
PyJWT==2.9.0
cryptography==43.0.1

# Stripe
stripe==10.12.0

# Utilities
python-multipart==0.0.12    # File uploads
python-dotenv==1.0.1
structlog==24.4.0            # Structured logging
tenacity==9.0.0              # Retry logic
slowapi==0.1.9               # Rate limiting (Redis-backed)

# Dev only
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2                # Test client
